+++ packages/Chem/src/widgets/chem-substructure-filter.ts
@@ -146,7 +146,7 @@ export class SubstructureFilter extends DG.Filter {
     this.optionsIcon = ui.icons.settings(() => {
       this.onShowOptionsChanged();
     });
-    $(this.optionsIcon).addClass('search-options-icon');
+    $(this.optionsIcon).addClass('chem-search-options-icon');
 
     this.sketcherDiv.append(this.sketcher.root);
     this.emptySketcherDiv.append(this.searchTypeInput.root);
@@ -404,7 +404,7 @@ export class SubstructureFilter extends DG.Filter {
   }
 
   async getFilterBitset(): Promise<BitArray> {
-    const smarts = !this.sketcher.sketcher?.isInitialized ? await this.sketcher.getSmarts() :
+    const smarts = this.sketcher.sketcher?.isInitialized ? await this.sketcher.getSmarts() :
       _convertMolNotation(this.currentMolfile, DG.chem.Notation.MolBlock, DG.chem.Notation.Smarts, getRdKitModule());
     return await chemSubstructureSearchLibrary(this.column!, this.currentMolfile, smarts!, false, false,
       this.searchType, this.similarityCutOff, this.fp);
@@ -435,7 +435,7 @@ export class SubstructureFilter extends DG.Filter {
         this.emptySketcherDiv.append(this.searchTypeInput.root);
         this.emptySketcherDiv.append(this.sketcherDiv);
         this.root.append(this.emptySketcherDiv);
-        this.removeChildIfExists(this.sketcher.root, this.optionsIcon, 'search-options-icon');
+        this.removeChildIfExists(this.sketcher.root, this.optionsIcon, 'chem-search-options-icon');
         if (this.searchType !== SubstructureSearchType.IS_SIMILAR)
           this.removeChildIfExists(this.searchOptionsDiv, this.similarityOptionsDiv, 'chem-filter-similarity-options');
       }
@@ -451,7 +451,7 @@ export class SubstructureFilter extends DG.Filter {
   onSearchTypeChanged() {
     this.searchType = this.searchTypeInput.value;
     if (this.searchType !== SubstructureSearchType.CONTAINS)
-      this.removeChildIfExists(this.sketcher.root, this.optionsIcon, 'search-options-icon');
+      this.removeChildIfExists(this.sketcher.root, this.optionsIcon, 'chem-search-options-icon');
     if (this.searchType === SubstructureSearchType.IS_SIMILAR)
       this.searchOptionsDiv.append(this.similarityOptionsDiv);
     else